requirements.txt
Flask
Flask-WTF
Flask-SQLAlchemy
SQLAlchemy
email-validator
Werkzeug
python-dotenv
pytest
